Fixed type in constructor

+++ 068_vector_ctor/vector.hpp
@@ -7,11 +7,11 @@ private:
   double x;
   double y;
 public:
-  Vector2D () : x(0),
-                y(0) {
+  Vector2D () : x(0.0),
+                y(0.0) {
   }
-  Vector2D (int init_x, int init_y) : x(init_x),
-                                      y(init_y) {
+  Vector2D (double init_x, double init_y) : x(init_x),
+                                            y(init_y) {
   }
   double getMagnitude () const;
   Vector2D operator+ (const Vector2D & rhs) const;
